move map to queue
