+++ src/queries/upload-queue.tsx
@@ -1,23 +1,7 @@
 import { createMutation, queryOptions, useQueryClient } from '@tanstack/solid-query'
 import { fetcher } from '~/api'
 import { makeAthenaCall } from '~/api/athena'
-import { AthenaCallResponse, AthenaOfflineQueueResponse, UploadFilesToUrlsRequest, UploadQueueItem } from '~/types'
-
-const mapOnlineQueueItems = (data: AthenaCallResponse<UploadQueueItem[]>) => data.result?.sort((a, b) => b.progress - a.progress) || []
-const mapOfflineQueueItems = (data: AthenaOfflineQueueResponse): UploadQueueItem[] =>
-  data
-    .filter((item) => item.method === 'uploadFilesToUrls')
-    .flatMap((item) =>
-      (item.params as UploadFilesToUrlsRequest).files_data.map((file) => ({
-        ...file,
-        path: file.fn,
-        created_at: 0,
-        current: false,
-        id: '',
-        progress: 0,
-        retry_count: 0,
-      })),
-    )
+import { AthenaOfflineQueueResponse, UploadQueueItem } from '~/types'
 
 export const uploadQueue = {
   prefix: ['upload_queue'],
@@ -28,7 +12,6 @@ export const uploadQueue = {
     queryOptions({
       queryKey: uploadQueue.onlineForDongle(dongleId),
       queryFn: () => makeAthenaCall<void, UploadQueueItem[]>(dongleId, 'listUploadQueue'),
-      select: mapOnlineQueueItems,
     }),
 
   offline: () => [...uploadQueue.prefix, 'offline'],
@@ -37,7 +20,6 @@ export const uploadQueue = {
     queryOptions({
       queryKey: uploadQueue.offlineForDongle(dongleId),
       queryFn: () => fetcher<AthenaOfflineQueueResponse>(`/v1/devices/${dongleId}/athena_offline_queue`),
-      select: mapOfflineQueueItems,
     }),
 
   cancelUpload: (dongleId: string) => {
